requirements.txt
streamlit
google-generativeai
langchain
langchain-community
chromadb
pypdf
